REF: extract LP interface as module `polytope.solvers`
so that it can be used in more than one module.

The new module is not hidden in order to allow the
user to select the default solver by setting a variable
in the global namespace of the module `solvers`.
